Update script.js
Changed minor aspects such as more numbers (1-12) and an extra message to feedback number of attempts left for user. 

Bug occurred in last iteration where the code would log a loss if the user guessed the correct number on the last attempt.
To fix this i added an extra if statement in handleGuess function to allow for a different message if this situation occurs.